feat(manager): Working tui and skeleton
Squashed and rebased to root.

This commit contains a lot of changes.
Ideally commit history will be reworked to start with the tui and
runner.
With all puzzle commits reworked and moved to a separate branch.

chore(tools): Improve and add tests to tools

fmt: Run formatting on all files