feat(manager): Add pre init for run types
Some problems require a different setup for the samples and input runs.
i.e. grid size, iteration count, etc.

These values are not part of the input, but are part of the problem
statement. A Pre function can now be defined which will be called with
the input file name. This allows some preliminary setup for the current
run.